backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
certifi>=2024.2.2
whitenoise>=6.6,<7.0
gunicorn>=21.2,<22.0
